Rename `utils2.py` to `metrics.py`
